fix(#11): 이미지 업데이트 시 default_user_image는 삭제되지 않게 수정

+++ src/users/controllers/user-image.controller.ts
@@ -22,7 +22,7 @@ export class UserImageController {
       const imageKey = imageUrlParts[imageUrlParts.length - 1]; // S3에 업로드된 이미지의 키
       const dbImageUrl = imageUrlParts[imageUrlParts.length - 2]; // 이미지 제공자 이름
       
-      if (dbImageUrl == 'ma6-main.s3.ap-northeast-2.amazonaws.com') {
+      if (dbImageUrl == 'ma6-main.s3.ap-northeast-2.amazonaws.com' && imageKey !== 'default_user_image.png') { // S3에 업로드된 이미지이고, 기본 이미지가 아닌 경우
         await this.s3Service.deleteImage(imageKey); // S3에 업로드된 기존 이미지 삭제
       }
